backend/openai_service/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
openai==0.27.0
python-dotenv==1.0.0 